VisualMessages moved to ChatUI project to solve #15 Introduction of .NET 5 blocked from strict dependency of ChatLib from graphics (WPF)
